requirements.txt
aniso8601==8.0.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
google-auth==1.8.1
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kubernetes==10.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
websocket-client==0.56.0
Werkzeug==0.16.0
